[spirv] Add a better tiled and vectorized convolution path
This commit tiles a 2-D convolution along the output height, width,
and channel dimensions. The tiled convolution is distributed to
workgroup of size 32x8x1, with each group containing 8x2x1
invocations. This is just the initial configuration for Mali to
get started. After tiling, we further vectorize the convolution
for each GPU invocation.