[python] Add a pre_passes field to CompilerOptions
We have an `extra_args` field that provides callers the flexibilty
to append additional command-line arguments. This makes it easier
to inject passes ahead of the main translation pipeline too as
the position of the passes matter. This can be used for the cases
like converting the model from fp32 to fp16 ahead of translation.