[Flow] Add pass for cropping the program
Opening a PR to see if this or something similar would be wanted in core
IREE. Essentially this enables cropping the input program at a specified
dispatch. We have been using this in SHARK-Runtime as a debugging tool.

For example, for certain backends and/or large models,
`--iree-flow-trace-dispatch-tensors` is far too slow and produces too
much output to easily iterate with so being able to quickly "binary
search" can enable much faster isolation of correctness issues. This
also has proven useful for isolating "soft" crashes/unwanted behavior
like codegen blowing up (happens sometimes in SPIR-V).

The current implementation here is a bit hacky (and assumes single
output everywhere) so feedback on a "correct" way to implement this
is welcome.