[metal] Support push constants via uniform buffer emulation
Push constants are expected to be directly embedded in the binding
tables and avoid redirections. In Metal argument encoder, we have
`constantDataAtIndex` allowing inlined constants. However, due to
the current limitation, we are unable to generate the suitable MSL
for such patterns. So we emulate push constants by emulating them
with normal uniform buffers.