[metal] Add builtin executables for polyfilling buffer fills
On macOS, the Metal API does not support buffer fills with
non-4-multiple buffer offsets and lengths for fills. It also
does not support 2/4-byte fill patterns. For such cases, we
need to use a kernel dispatch to polyfill it.

Three kernels are provided for different cases, where we can
optimize for 16-byte or 4-byte aligned cases a bit better.