[cuda] Use command segments in graph command buffer impl
This commit switches the graph command buffer to use command
segments to handle recording. This is a prelimiary step for
better dependency analysis and handling. The reasons are:

In a CUDA graph, buffer management and kernel launches are
represented as graph nodes. Dependencies are represented by
graph edges. IREE's HAL follows the Vulkan command buffer
recording model, which "linearizes" the original graph.
So we have a mismatch here. Implementing IREE's HAL using
CUDA graph would require rediscover the graph node dependencies
from the linear chain of command buffer commands; it means
looking at both previous and next commands sometimes.

Due to these reasons, it's beneficial to have a complete view
of the full command buffer and extra flexibility during
recording, in order to fixup past commands, or inspect
past/future commands.

Note that this does not improve the descriptor and push constant
handling yet and also does not really change the node
serialization. Those are addressed in future commits.